% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_length_cca.R
\name{compare_length_cca}
\alias{compare_length_cca}
\title{Compare length difference between inside and outside CCA}
\usage{
compare_length_cca(
  dir,
  data,
  file = "hkl_cca_comparison",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{dir}{directory to save plost}

\item{data}{dataframe name}

\item{file}{specialized file name to use for the figure}

\item{xlim}{custom xlim input, otherwise dynamically determined}

\item{ylim}{custom ylim input, otherwise dynamically determined}
}
\value{
plots
}
\description{
Compare length difference between inside and outside CCA
}
\author{
Chantel Wetzel
}
