% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_frame.R
\name{create_data_frame}
\alias{create_data_frame}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
create_data_frame(data_list)
}
\arguments{
\item{list}{of data sets}
}
\value{
A data frame
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
