% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hist_plot.R
\name{data_hist}
\alias{data_hist}
\title{Create histogram of length or age samples by area and by fleet}
\usage{
data_hist(
  dir,
  data,
  data_type = "Length",
  group_column = "State_Areas",
  fleet_column = "Fleet",
  ymax = NULL,
  do_abline = TRUE
)
}
\arguments{
\item{dir}{directory of where save save plots}

\item{data}{data frame created by the create_data_frame function}

\item{data_type}{plot either the length or age data (default "Length")}

\item{group_column}{column name from data frame to group the data for plotting (e.g., State for
to see the data by state)}

\item{fleet_column}{column name from data frame to group for fleet structure}

\item{ymax}{vector of ymax values that should be equal in to unique group_column values. Default
input is NULL which sets the ymax to 0.15 for all group_column values.}

\item{do_abline}{}
}
\value{
Nothing - plots
}
\description{
Create histogram of length or age samples by area and by fleet
}
\author{
Chantel Wetzel
}
