% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_length_age.R
\name{estimate_length_age}
\alias{estimate_length_age}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
estimate_length_age(data, grouping = "all", linf = NULL, l0 = NULL, k = NULL)
}
\arguments{
\item{grouping}{not implemented (options: all, sex, source, state)}

\item{linf}{L infinity (L2) for the user to specifiy if desired.}

\item{l0}{Initial length (L1) for the user to specifiy if desired.}

\item{k}{von bert growth rate for the user to specifiy if desired.}

\item{list}{of data}
}
\value{
data frame with estimated length-at-age
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
