% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_length_weight.R
\name{estimate_length_weight}
\alias{estimate_length_weight}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
estimate_length_weight(data, grouping = "all")
}
\arguments{
\item{grouping}{options: all, source, state}

\item{list}{of data}
}
\value{
data frame with estimated length-at-weight
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
