% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_age_plot.R
\name{length_age_plot}
\alias{length_age_plot}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
length_age_plot(
  dir,
  data,
  splits = NA,
  nm_append = NULL,
  ests = NULL,
  plots = 1:4
)
}
\arguments{
\item{dir}{directory of where save save plots}

\item{data}{data frame created by the create_data_frame function}

\item{splits}{not implemented}

\item{nm_append}{text to append to the plot name}

\item{ests}{list of estimates created by the estimate_length_weight function}

\item{plots}{allows for a subset of plots to be selected
plot 1: Length-at-sge by from all sources and by each one for sexes combined
plot 2: Length-at-age by sex from all sources combined and by each one
plot 3: Length-at-age from each state with sexes combined     
plot 4: Length-at-age by sex from each state}
}
\value{
Nothing - plots
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
