% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_by_depth_plot.R
\name{length_by_depth_plot}
\alias{length_by_depth_plot}
\title{Compare length frequencies by data set}
\usage{
length_by_depth_plot(dir, data, xlim = NULL, ylim = NULL)
}
\arguments{
\item{dir}{directory to save the plots}

\item{data}{data frame created by create_data_frame}

\item{xlim}{custom xlim input, otherwise dynamically determined}

\item{ylim}{custom ylim input, otherwise dynamically determined}
}
\value{
A data frame
}
\description{
Compare length frequencies by data set
}
\author{
Chantel Wetzel
}
