% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight_plot.R
\name{length_weight_plot}
\alias{length_weight_plot}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
length_weight_plot(dir, data, splits = NA, nm_append = NULL, ests = NULL)
}
\arguments{
\item{dir}{directory of where save save plots}

\item{data}{data frame created by the create_data_frame function}

\item{splits}{not implemented}

\item{nm_append}{text to append to the plot name}

\item{ests}{list of estimates created by the estimate_length_weight function}
}
\value{
Nothing - plots
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
