% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacfin_fleets.R
\name{pacfin_fleets}
\alias{pacfin_fleets}
\title{Create fleet naming for PacFIN}
\usage{
pacfin_fleets(data, fleet_grouping, fleet_names, fleet_column_name)
}
\arguments{
\item{data}{read in PacFIN data}

\item{fleet_grouping}{list of area names in data source}

\item{fleet_names}{user area names}

\item{column_name}{name to do the grouping with (default is based on recfin)}
}
\value{
A data frame
}
\description{
Create fleet naming for PacFIN
}
\author{
Chantel Wetzel
}
