% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngfun.R
\name{pngfun}
\alias{pngfun}
\title{Function to simplify the creation of png files}
\usage{
pngfun(wd, file, w = 7, h = 7, pt = 12)
}
\arguments{
\item{wd}{Directory to save files}

\item{file}{File name}

\item{w}{width of png figure}

\item{h}{height of png figure}

\item{pt}{Pixel size in figure}
}
\value{

}
\description{
Function to simplify the creation of png files
}
\author{
Chantel Wetzel
}
