% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_budrick_recfin.R
\name{rename_budrick_recfin}
\alias{rename_budrick_recfin}
\title{Modify California recreational data that does not match RecFIN column labels
for use by data exploration tools}
\usage{
rename_budrick_recfin(data)
}
\arguments{
\item{data}{pass read in data provided by John Budrick for CA recreational data}
}
\value{
A data frame
}
\description{
Modify California recreational data that does not match RecFIN column labels
for use by data exploration tools
}
\author{
Brian Langseth and Chantel Wetzel
}
