% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_hook_and_line.R
\name{rename_hook_and_line}
\alias{rename_hook_and_line}
\title{Create standarized field for hook & line data}
\usage{
rename_hook_and_line(data, survey_name = NULL)
}
\arguments{
\item{data}{read in hook and line data}

\item{survey_name}{}
}
\value{
A data frame
}
\description{
Create standarized field for hook & line data
}
\author{
Chantel Wetzel
}
