% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_mrfss.R
\name{rename_mrfss}
\alias{rename_mrfss}
\title{Create standarized field for RecFIN data}
\usage{
rename_mrfss(
  data,
  len_col = "T_LEN",
  len_divide = 10,
  area_grouping = NULL,
  area_names = NULL,
  area_column_name = "DIST",
  mode_grouping = NULL,
  mode_names = NULL,
  mode_column_name = NULL
)
}
\arguments{
\item{data}{read in RecFIN data}

\item{len_col}{name of the length column to use}

\item{len_divide}{value to convert between mm and cm}

\item{area_grouping}{list of area names in data source}

\item{area_names}{user area names}

\item{column_name}{area field that can be used to specify catch area}
}
\value{
A data frame
}
\description{
Create standarized field for RecFIN data
}
\author{
Chantel Wetzel
}
