% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_pacfin.R
\name{rename_pacfin}
\alias{rename_pacfin}
\title{Create standarized cleaning and renaming PacFIN data.
This should be folded into the cleanPacFIN function.}
\usage{
rename_pacfin(data, area_grouping = NULL, area_names = NULL)
}
\arguments{
\item{data}{read in pacfin data file}

\item{area_grouping}{list of area names in data source}

\item{area_names}{user area names}
}
\value{
A data frame
}
\description{
Create standarized cleaning and renaming PacFIN data.
This should be folded into the cleanPacFIN function.
}
\author{
Chantel Wetzel
}
