% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_pacfin_newVersion.R
\name{rename_pacfin_newVersion}
\alias{rename_pacfin_newVersion}
\title{Create standarized cleaning and renaming PacFIN data.
Used with updated version of PacFIN.Utilities 0.0.1.9999
as of Feb. 2021}
\usage{
rename_pacfin_newVersion(
  data,
  area_grouping = NULL,
  area_names = NULL,
  fleet_grouping = NULL,
  fleet_names = NULL,
  fleet_column_name = "COND"
)
}
\arguments{
\item{data}{read in pacfin data file}

\item{area_grouping}{list of grouped area names in data source}

\item{area_names}{user area names}

\item{fleet_grouping}{list of grouped fleet names in data source}

\item{fleet_names}{user fleet names}

\item{fleet_column_name}{name of the column to group fleets by in the PacFIN data}
}
\value{
A data frame
}
\description{
Create standarized cleaning and renaming PacFIN data.
Used with updated version of PacFIN.Utilities 0.0.1.9999
as of Feb. 2021
}
\author{
Chantel Wetzel, Brian Langseth
}
