% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_recfin.R
\name{rename_recfin}
\alias{rename_recfin}
\title{Create standarized field for RecFIN data}
\usage{
rename_recfin(data, area_grouping = NULL, area_names = NULL)
}
\arguments{
\item{data}{read in RecFIN data}

\item{area_grouping}{list of area names in data source}

\item{area_names}{user area names}
}
\value{
A data frame
}
\description{
Create standarized field for RecFIN data
}
\author{
Chantel Wetzel
}
