% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_data.R
\name{rename_survey_data}
\alias{rename_survey_data}
\title{Create standarized field for survey data}
\usage{
rename_survey_data(data)
}
\arguments{
\item{data}{read in biological survey data}
}
\value{
A data frame
}
\description{
Create standarized field for survey data
}
\author{
Chantel Wetzel
}
