% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_wa_recfin.R
\name{rename_wa_recfin}
\alias{rename_wa_recfin}
\title{Modify Washington recreational data that does not match RecFIN column labels
for use by data exploration tools}
\usage{
rename_wa_recfin(data)
}
\arguments{
\item{data}{pass read in data provided by John Budrick for CA recreational data}
}
\value{
A data frame
}
\description{
Modify Washington recreational data that does not match RecFIN column labels
for use by data exploration tools
}
\author{
Brian Langseth and Chantel Wetzel
}
