% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize all the data to evaluate the samples by source,
year, and biological measurement (length, weight, age)}
\usage{
summarize_data(dir = NULL, data)
}
\arguments{
\item{dir}{directory location to save output}

\item{data}{data frame created by the create_data_frame function}
}
\value{
list
}
\description{
Summarize all the data to evaluate the samples by source,
year, and biological measurement (length, weight, age)
}
\author{
Chantel Wetzel
}
