% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_functions.R
\name{vb_fn}
\alias{vb_fn}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
vb_fn(age, Linf, L0, k)
}
\arguments{
\item{age}{vector of ages to predict length for (e.g., age = 1:60)}

\item{Linf}{maximum length parameter}

\item{L0}{initial length parameter when growth changes}

\item{k}{von Bertenlaffy growth parameter}
}
\value{
A vector of
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
