% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_functions.R
\name{vb_fn}
\alias{vb_fn}
\title{Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources}
\usage{
vb_fn(age_vec, Linf, k, t0)
}
\arguments{
\item{t0}{}
}
\value{
A vector of
}
\description{
Create function to take data from multiple sources and create a 
single data frame that can be used for biological comparisons acroos
data sources
}
\author{
Chantel Wetzel
}
